feat: finish: memory_reserve()

+++ lab4/kernel/include/string.h
@@ -1,8 +1,8 @@
-#include "stddef.h"
+#include "stdint.h"
 
 size_t strlen(const char *str);
 int strcmp(const char *p1, const char *p2);
-int strncmp(const char *s1, const char *s2, unsigned long long n);
+int strncmp(const char *s1, const char *s2, size_t n);
 char *strcpy(char *dest, const char *src);
 int atoi(char* str);
 void *memcpy(void *dest, const void *src, size_t count);
